remove counter interface and quota
