--- src/RateLimiter/CounterInterface.php
@@ -1,12 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Yiisoft\Yii\Web\RateLimiter;
-
-interface CounterInterface
-{
-    public function limitIsReached(): bool;
-
-    public function setId(string $id): void;
-}
